refactor(web): move useCanManageGroups to the permissions feature lib
web/CLAUDE.md keeps web/src/hooks/ for general-purpose hooks with no feature
home; a tier gate on group administration is domain policy and belongs in the
feature lib.

lib/permissions/hooks.ts rather than lib/users/: it already holds
usePermissionAuthority, so both "what authority does this caller have here"
questions now read together, and related tier behaviour is harder to change in
only one place.

useGroups.ts returns to its original contents and leaves the diff.
