+++ web/src/lib/permissions/hooks.ts
@@ -1,6 +1,8 @@
 import { useUser } from "@/providers/UserProvider";
 import { hasPermission } from "@/lib/permissions";
 import { Permission } from "@/lib/types";
+import { useTierAtLeast } from "@/hooks/useTierAtLeast";
+import { Tier } from "@/lib/settings/types";
 
 export interface PermissionAuthority {
   /** Holds the permission outright, or is an admin — unrestricted org-wide. */
@@ -31,4 +33,13 @@ export function usePermissionAuthority(
     isScopedManager:
       !isGlobalHolder && hasPermission(adminCapabilities, permission),
   };
+}
+
+/**
+ * Mirrors the backend BUSINESS gate on `/manage/admin/user-group`. Gate edit
+ * affordances on this, not `settings.enterprise` — that only means "EE build",
+ * and below Business the endpoint 402s.
+ */
+export function useCanManageGroups(): boolean {
+  return useTierAtLeast(Tier.BUSINESS);
 }
